Docker

+++ models/models.js
@@ -1,5 +1,6 @@
 import {sequelize} from '../db.js';
 import { DataTypes} from 'sequelize';
+import {} from 'dotenv/config';
 
 export const TypeContacts = sequelize.define('type_contacts', {
   id: {type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true},
@@ -28,7 +29,7 @@ export const TypeProject = sequelize.define('type_project', {
 });
 
 export const User = sequelize.define('user', {
-  id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
+  id: {type: DataTypes.BIGINT, primaryKey: true, autoIncrement: true},
   email: {type: DataTypes.STRING, unique: true},
   password: {type: DataTypes.STRING},
   role: {type: DataTypes.STRING, defaultValue: 'USER'},
